Ahh, this brings back memories.
This is one of the first games I programmed. The first one I ever wrote
in Python. It was shortly after I had learned about Pentago but at the
time I didn't have the money to buy it so I decided to write this so
that I could try it out and play against people.